requirements.txt
xgboost==1.3.0.post0
streamlit==0.82.0
pandas==1.2.4
Pillow==8.3.1
